+++ lib/src/database/database_utils.dart
@@ -18,7 +18,8 @@ abstract class DatabaseUtils {
   /// See: https://developer.android.com/reference/android/database/DatabaseUtils.html#dumpCurrentRow(android.database.Cursor)
   static Future<void> dumpCurrentRow(final Cursor cursor) async {
     assert(_platform.isAndroid);
-    // TODO
+
+    print(await dumpCurrentRowToString(cursor));
   }
 
   /// Prints the contents of a [Cursor]'s current row to a string.
@@ -26,7 +27,43 @@ abstract class DatabaseUtils {
   /// See: https://developer.android.com/reference/android/database/DatabaseUtils.html#dumpCurrentRowToString(android.database.Cursor)
   static Future<String> dumpCurrentRowToString(final Cursor cursor) async {
     assert(_platform.isAndroid);
-    return null; // TODO
+
+    final buffer = StringBuffer();
+    dumpCurrentRowToStringBuffer(cursor, buffer);
+    return buffer.toString();
+  }
+
+  /// Prints the contents of a [Cursor]'s current row to a string buffer.
+  ///
+  /// See: https://developer.android.com/reference/android/database/DatabaseUtils.html#dumpCurrentRowToString(android.database.Cursor)
+  static Future<void> dumpCurrentRowToStringBuffer(final Cursor cursor, final StringBuffer buffer) async {
+    assert(_platform.isAndroid);
+
+    buffer.write("[");
+    for (int columnIndex = 0; columnIndex < cursor.getColumnCount(); columnIndex++) {
+      if (columnIndex > 0) buffer.write(", ");
+      switch (cursor.getType(columnIndex)) {
+        case Cursor.FIELD_TYPE_NULL:
+          buffer.write("null");
+          break;
+        case Cursor.FIELD_TYPE_INTEGER:
+          buffer.write(cursor.getLong(columnIndex));
+          break;
+        case Cursor.FIELD_TYPE_FLOAT:
+          buffer.write(cursor.getDouble(columnIndex));
+          break;
+        case Cursor.FIELD_TYPE_STRING:
+          buffer.write(cursor.getString(columnIndex));
+          break;
+        case Cursor.FIELD_TYPE_BLOB:
+          buffer.write(cursor.getBlob(columnIndex));
+          break;
+        default:
+          assert(false); // unreachable
+          throw AssertionError();
+      }
+    }
+    buffer.write("]");
   }
 
   /// Prints the contents of a [Cursor] to standard output.
@@ -36,7 +73,12 @@ abstract class DatabaseUtils {
   /// See: https://developer.android.com/reference/android/database/DatabaseUtils.html#dumpCursor(android.database.Cursor)
   static Future<void> dumpCursor(final Cursor cursor) async {
     assert(_platform.isAndroid);
-    // TODO
+
+    final position = cursor.getPosition();
+    while (cursor.moveToNext()) {
+      print(await dumpCurrentRowToString(cursor));
+    }
+    cursor.moveToPosition(position);
   }
 
   /// Prints the contents of a [Cursor] to a string.
@@ -46,6 +88,16 @@ abstract class DatabaseUtils {
   /// See: https://developer.android.com/reference/android/database/DatabaseUtils.html#dumpCursorToString(android.database.Cursor)
   static Future<String> dumpCursorToString(final Cursor cursor) async {
     assert(_platform.isAndroid);
-    return null; // TODO
+
+    final buffer = StringBuffer();
+
+    final position = cursor.getPosition();
+    while (cursor.moveToNext()) {
+      await dumpCurrentRowToStringBuffer(cursor, buffer);
+      buffer.writeln();
+    }
+    cursor.moveToPosition(position);
+
+    return buffer.toString();
   }
 }